fix linux build * gcc 9.4 doesn't support non-ascii identifiers (ε 🡒 epsilon, π 🡒 pi) * media_query.cpp: '<=' 🡒 u'⩽' to fix gcc warning, media_query.cpp now has UTF-8 BOM for Visual Studio, alternatively /utf-8 compiler switch can be used

+++ include/litehtml/html.h
@@ -51,7 +51,7 @@ namespace litehtml
 	{
 		static T invalid_item; // T's default constructor must create invalid item
 		if (index < 0) index += (int)vec.size();
-		return index >= 0 && index < vec.size() ? vec[index] : invalid_item;
+		return index >= 0 && index < (int)vec.size() ? vec[index] : invalid_item;
 	}
 	template<typename T>
 	vector<T> slice(const vector<T>& vec, int index, int count = -1)
@@ -64,7 +64,7 @@ namespace litehtml
 	{
 		if (index < 0) index += (int)vec.size();
 
-		if (!(index >= 0 && index < vec.size()))
+		if (!(index >= 0 && index < (int)vec.size()))
 			return;
 
 		count = min(count, (int)vec.size() - index);
